Move generic host headers to include
