+++ include/generic_host/IHostedService.hpp
