+++ include/generic_host/ILogger.hpp
